Restructure translator package
