+++ main/src/main/java/tw/firemaples/onscreenocr/floatings/manager/FloatingStateManager.kt
@@ -26,7 +26,7 @@ import tw.firemaples.onscreenocr.recognition.RecognitionResult
 import tw.firemaples.onscreenocr.recognition.TextRecognitionProviderType
 import tw.firemaples.onscreenocr.recognition.TextRecognizer
 import tw.firemaples.onscreenocr.screenshot.ScreenExtractor
-import tw.firemaples.onscreenocr.translator.MicrosoftAzureTranslator
+import tw.firemaples.onscreenocr.translator.azure.MicrosoftAzureTranslator
 import tw.firemaples.onscreenocr.translator.TranslationProviderType
 import tw.firemaples.onscreenocr.translator.TranslationResult
 import tw.firemaples.onscreenocr.translator.Translator
